chore(files): moving json catalog to dedicated folders
