+++ packages/backend/src/managers/catalogManager.spec.ts
@@ -19,8 +19,8 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
 
 import { beforeEach, describe, expect, test, vi } from 'vitest';
-import content from '../ai-test.json';
-import userContent from '../ai-user-test.json';
+import content from '../tests/ai-test.json';
+import userContent from '../tests/ai-user-test.json';
 import { type Webview, EventEmitter } from '@podman-desktop/api';
 import { CatalogManager } from './catalogManager';
